+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `TiVec::extend_from_within_corrected` method, a corrected variant of
+  `TiVec::extend_from_within` method.
 
 ## [3.4.0] - 2025-12-05
 ### Added